+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -34,7 +34,8 @@
 			<beans:bean class="org.springframework.web.socket.server.support.HttpSessionHandshakeInterceptor"/>
 		</websocket:handshake-interceptors>
 	</websocket:handlers>
-	<beans:bean id="myHandler" class="com.tis.websocket.WebSocket"/>
+	<beans:bean id="myHandler" class="com.tis.websocket.WebSocketHandler">
+	</beans:bean>
 	
 	<context:component-scan
 		base-package="com.tis.ridinghan" />
